requirements.txt
flask
google-generativeai
python-dotenv
feedparser
beautifulsoup4
requests
gunicorn
